cos/ecn.rs: colocate 15 ECN-marking tests with production (#984 P3 Phase 2c)
Move 15 tests targeting ethernet_l3, mark_ecn_ce_ipv4,
mark_ecn_ce_ipv6, and maybe_mark_ecn_ce from tx/mod.rs::tests to
a new mod tests in cos/ecn.rs.

Adds explicit use crate::afxdp::PROTO_TCP for the test scope.

Mechanical: 15 tests, 322 lines added; 308 removed from tx/mod.rs.
cargo test --bins 865/0/2.
